refactor: restructure application flow and add connect screen
This commit refactors the application's structure to introduce a dedicated connection screen before the main application view.

- Renames `MainAppScreen` to `HomeScreen` and repurposes it to hold the bottom navigation bar and the main feature screens (Control, Graph, Settings).
- Introduces a new `ConnectScreen` which handles Bluetooth permission requests, enabling Bluetooth, and connecting to the specified ARDUINOBT device.
- Moves the logic for checking Bluetooth status, requesting permissions, and initiating the connection from the previous `HomeScreen` (now `ConnectScreen`).
- Updates the navigation flow to navigate from `ConnectScreen` to `HomeScreen` upon successful Bluetooth connection.
- Creates a `MainAppBar` widget for consistent branding in the connection screen.
- Adjusts imports in `ControlCubit`, `GraphCubit`, and `SettingsCubit` to reflect the moved `BluetoothService` utility file.
