--- lib/features/home/presentation/view/screen/main_app_screen.dart
@@ -1,110 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:fluttertoast/fluttertoast.dart';
-
-import '../../../../../core/bluetooth_service.dart';
-import '../../../../control/presentation/managers/control_cubit.dart';
-import '../../../../control/presentation/view/control_screen.dart';
-import '../../../../graph/presentation/managers/graph_cubit.dart';
-import '../../../../graph/presentation/view/graph_screen.dart';
-import '../../../../settings/presentation/managers/settings_cubit.dart';
-import '../../../../settings/presentation/view/settings_screen.dart';
-
-class MainAppScreen extends StatefulWidget {
-  final BluetoothService bluetoothService;
-
-  const MainAppScreen({super.key, required this.bluetoothService});
-
-  @override
-  State<MainAppScreen> createState() => _MainAppScreenState();
-}
-
-class _MainAppScreenState extends State<MainAppScreen> {
-  int _selectedIndex = 0;
-  late List<Widget> _screens;
-
-  @override
-  void initState() {
-    super.initState();
-    _screens = [ControlScreen(), GraphScreen(), SettingsScreen()];
-  }
-
-  void _onItemTapped(int index) {
-    setState(() {
-      _selectedIndex = index;
-    });
-  }
-
-  Future<void> _handleLogout() async {
-    try {
-      await widget.bluetoothService.disconnect();
-      Fluttertoast.showToast(
-        msg: "Disconnected",
-        backgroundColor: Colors.orange,
-      );
-    } catch (e) {
-      Fluttertoast.showToast(
-        msg: "Error disconnecting: $e",
-        backgroundColor: Colors.red,
-      );
-    } finally {
-      Navigator.pushReplacementNamed(context, '/home');
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return MultiBlocProvider(
-      providers: [
-        BlocProvider(create: (_) => ControlCubit(widget.bluetoothService)),
-        BlocProvider(create: (_) => SettingsCubit(widget.bluetoothService)),
-        BlocProvider(create: (_) => GraphCubit(widget.bluetoothService)),
-      ],
-      child: Scaffold(
-        appBar: AppBar(
-          title: const Text("CoolFan Main App"),
-          backgroundColor: Colors.blueAccent,
-          actions: [
-            IconButton(
-              icon: const Icon(Icons.logout),
-              tooltip: 'Disconnect',
-              onPressed: _handleLogout,
-            ),
-          ],
-        ),
-        body: Container(
-          decoration: BoxDecoration(
-            gradient: LinearGradient(
-              colors: [
-                Colors.blueGrey[900]!,
-                Colors.blueAccent.withOpacity(0.3),
-              ],
-              begin: Alignment.topCenter,
-              end: Alignment.bottomCenter,
-            ),
-          ),
-          child: _screens[_selectedIndex],
-        ),
-        bottomNavigationBar: BottomNavigationBar(
-          items: const [
-            BottomNavigationBarItem(
-              icon: Icon(Icons.control_camera),
-              label: 'Control',
-            ),
-            BottomNavigationBarItem(
-              icon: Icon(Icons.show_chart),
-              label: 'Graph',
-            ),
-            BottomNavigationBarItem(
-              icon: Icon(Icons.settings),
-              label: 'Settings',
-            ),
-          ],
-          currentIndex: _selectedIndex,
-          selectedItemColor: Colors.blueAccent,
-          onTap: _onItemTapped,
-        ),
-      ),
-    );
-  }
-}
