+++ lib/features/home/presentation/view/widgets/main_app_bar.dart
@@ -1,0 +1,35 @@
+import 'package:flutter/material.dart';
+
+class MainAppBar extends StatelessWidget {
+  const MainAppBar({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return AppBar(
+      elevation: 0,
+      title: Row(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          const Text(
+            "Cryo",
+            style: TextStyle(
+              fontSize: 28,
+              fontWeight: FontWeight.bold,
+              color: Colors.black,
+            ),
+          ),
+          const SizedBox(width: 8),
+          const Text(
+            "Control",
+            style: TextStyle(
+              fontSize: 28,
+              fontWeight: FontWeight.bold,
+              color: Colors.blueAccent,
+            ),
+          ),
+        ],
+      ),
+      backgroundColor: Colors.white,
+    );
+  }
+}
