+++ lib/features/settings/presentation/managers/settings_cubit.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
-import '../../../../core/bluetooth_service.dart';
+import '../../../../core/utils/bluetooth_service.dart';
 
 class SettingsCubit extends Cubit<bool> {
   final BluetoothService bluetoothService;
